fix(document): avoid server extractor imports in client bundle
